refactor(web worker): swap out service worker for web worker - web worker is created directly with Worker API - web worker registers callback for messages against itself - rename various functions - update docs
